I have implemented the FusedLocationApi
