+++ app/src/main/AndroidManifest.xml
@@ -4,7 +4,7 @@
 
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
-    <uses-permission android:name="android.permission.CALL_PHONE"></uses-permission>
+    <uses-permission android:name="android.permission.CALL_PHONE" />
 
     <application
         android:allowBackup="true"
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         android:screenOrientation="portrait">
-        <activity android:name=".HomeActivity"
+        <activity android:name=".activities.HomeActivity"
             android:screenOrientation="portrait">
 
             <intent-filter>
@@ -36,19 +36,19 @@
 
         <activity
             android:screenOrientation="portrait"
-            android:name=".NavigateActivity"
+            android:name=".activities.NavigateActivity"
             android:label="@string/title_activity_navigate"></activity>
         <activity
             android:screenOrientation="portrait"
-            android:name=".UserActivity"
+            android:name=".activities.UserActivity"
             android:label="@string/title_activity_user"></activity>
         <activity
         android:screenOrientation="portrait"
-        android:name=".RoutesActivity"
+        android:name=".activities.RoutesActivity"
         android:label="@string/title_activity_routes"></activity>
         <activity
             android:screenOrientation="portrait"
-            android:name=".CarerActivity"
+            android:name=".activities.CarerActivity"
             android:label="@string/title_activity_carer"></activity>
     </application>
