+++ app/src/main/java/com/example/hamsajama/wayfindandorient/activities/CarerActivity.java
@@ -1,7 +1,9 @@
-package com.example.hamsajama.wayfindandorient;
+package com.example.hamsajama.wayfindandorient.activities;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+
+import com.example.hamsajama.wayfindandorient.R;
 
 /**
  * Created by hamsajama on 12/08/2016.
