+++ app/src/main/java/com/example/hamsajama/wayfindandorient/activities/UserActivity.java
@@ -1,4 +1,4 @@
-package com.example.hamsajama.wayfindandorient;
+package com.example.hamsajama.wayfindandorient.activities;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -6,6 +6,7 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 
+import com.example.hamsajama.wayfindandorient.R;
 import com.example.hamsajama.wayfindandorient.motionUtil.MotionDection;
 
 public class UserActivity extends AppCompatActivity  {
