+++ app/src/main/res/layout-v21/activity_main.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".UserActivity"
+    tools:context=".activities.UserActivity"
     android:background="@color/colorPrimary">
 
     <Button
