+++ app/src/main/res/layout/activity_navigate.xml
@@ -15,7 +15,7 @@
         android:name="com.google.android.gms.maps.SupportMapFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".NavigateActivity"
+        tools:context=".activities.NavigateActivity"
         android:layout_gravity="left|center_vertical" />
 </FrameLayout>
